Fix a bug where the target/logger were holding on to reference to an old driver, which would fail if the session was relaunched and a new driver was created.
RemoteProxy doesn’t actually require the driver, it just requires an
object that responds to #execute_script, so I’ve redefined it’s 
collaborator as an “executor”. The iOS driver can act as the executor
and delegate to the driver of it’s current session. That way nothing
holds on to a reference to the web driver other than the Session.